requirements.txt
fastapi~=0.115.6
pydantic-settings~=2.7.0
SQLAlchemy~=2.0.36
passlib~=1.7.4
python-jose~=3.3.0
pydantic~=2.10.3